chore: refactor rule pushdown filter to join
