+++ src/query/sql/src/planner/optimizer/rule/rewrite/rewrite_util/mod.rs
@@ -1,0 +1,22 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod extract_or_predicates;
+mod mark_join_to_semi_join;
+mod outer_join_to_inner_join;
+
+pub use extract_or_predicates::rewrite_predicates;
+pub use mark_join_to_semi_join::convert_mark_to_semi_join;
+pub use outer_join_to_inner_join::convert_outer_to_inner_join;
+pub use outer_join_to_inner_join::remove_nullable;
